bringing code from main

+++ app/routes.py
@@ -1,11 +1,20 @@
+'''Contains the routes of the application.
+
+This module handles all the routes and logic of the FlaskApp. 
+This file should be imported in __init__.py to initialize all methods/routes of the FlaskApp.
+
+'''
+
+
 from math import ceil
-from flask import render_template, session, redirect, request, flash, url_for, get_flashed_messages
+from flask import render_template, redirect, request, flash, url_for
 from app import myapp, db
 from app.models import AuctionItem, Item,User 
 from flask_login import login_user, logout_user, login_required, current_user
-from app.forms import BidButton, register, LoginForm, SearchForm, purchaseItemForm, addToCart, 
-User, ListItemForm, CompareItemButton
-from app.forms import register, LoginForm, SearchForm, purchaseItemForm, addToCart, deleteUser, ListItemForm, CompareItemButton, deleteFromCart, 
+
+from app.forms import BidButton, register, LoginForm, SearchForm, purchaseItemForm, addToCart, deleteUser, ListItemForm, CompareItemButton
+from app.forms import register, LoginForm, SearchForm, purchaseItemForm, addToCart, deleteUser, ListItemForm, CompareItemButton, deleteFromCart, SellerPageForm, changePasssword
+ 
 from random import choice
 from datetime import datetime, timedelta
 
@@ -14,23 +23,18 @@ from datetime import datetime, timedelta
 @myapp.route('/')
 @myapp.route('/home', methods=['GET'])
 def home():
-    form = SearchForm()
-    suggestions = ['Bananas', 'iPad', 'Gaming Laptop']
-    return render_template('home.html', form=form, suggestion=str(choice(suggestions)) + '...')
-# @login_required
-# @app.route('/sellerspage')
-# def seller_page():
-# items= Item.query.filter_by( username= form.username.data).first()
-# return render_template('sellerspage.html', items=items)
-
+    '''The home page.'''
+    return render_template('home.html')
 
 
 
 @myapp.route('/login', methods= ['POST','GET'])
 def login():
+    '''The login page.'''
+    
     form = LoginForm()
     if form.validate_on_submit():
-        attempted_user = User.query.filter_by(username=form.username.data).first()
+        attempted_user : User = User.query.filter_by(username=form.username.data).first()
         if attempted_user and attempted_user.check_password_correction(
                 attempted_password=form.password.data):
             login_user(attempted_user)
@@ -65,9 +69,13 @@ def list_item():
             new_item = AuctionItem(name = form.item_name.data, price = form.item_price.data, 
                         picture = '', description = form.item_description.data, 
                         Owner = None, auction_end = auction_end, type = 'auction')
+            
         else:
             new_item = Item(name = form.item_name.data, price = form.item_price.data, 
                         picture = '', description = form.item_description.data, Owner = None, type = 'sale')
+
+
+        new_item.set_seller(current_user)             
         db.session.add(new_item)
         db.session.commit()
     return render_template('list.html', form=form)
@@ -119,10 +127,7 @@ def market():
                 flash(f"Congratulations! You purchased {p_item_object.name} for {p_item_object.price}$", category='success')
             else:
                 flash(f"Unfortunately, you don't have enough money to purchase {p_item_object.name}!", category='danger')
-        return redirect(url_for('market'))
 
-    add_to_cart = addToCart()
-    if request.method == "POST":
 
         cart_item = request.form.get('cart_item')
         cart_object = Item.query.filter_by(id = cart_item).first()
@@ -132,6 +137,12 @@ def market():
         else:
             flash("no cart")
         return redirect(url_for('market'))
+        
+
+
+
+
+
 
     if request.method == "GET":
         items = Item.query.filter_by(Owner=None)
@@ -189,7 +200,7 @@ def cart():
             else:
                 flash(f"Unfortunately, you don't have enough money to purchase {p_item_object.name}!", category='danger')
         remove_item = request.form.get('delete_item')
-        item_to_remove = Item.query.filter_by(id=remove_item).first()
+        item_to_remove = Item.query.get(remove_item)
         if item_to_remove:
             item_to_remove.remove_from_cart()
             
@@ -201,25 +212,42 @@ def cart():
         items = Item.query.filter_by(cart= current_user.id)
         return render_template('cart.html', items=items, checkout_form = checkout_form, delete_from_cart = delete_from_cart)
 
+
+
+
+
+
 @myapp.route('/profile', methods = ['POST', 'GET'])
 def profile():
 
     delete_user = deleteUser()
-    
+    form_change_password = changePasssword()
+
     if request.method == "POST":
-        db.session.delete(current_user)
-        db.session.commit()
-        flash("User deleted successfully")
-        return redirect(url_for('login'))
+        if form_change_password.validate_on_submit():
+            new_password = form_change_password.new_password.data
+            current_user.change_password(new_password)
+
+
+            
+        elif delete_user.validate_on_submit():
+            db.session.delete(current_user)
+            db.session.commit()
+            flash("User deleted successfully")
+            return redirect(url_for('login'))
+            
+
+
+
         
 
-    return render_template('profile.html', delete_user = delete_user)
+    return render_template('profile.html', delete_user = delete_user, form_change_password = form_change_password)
 
 @myapp.route('/compare', methods = ['POST'])
 def compare():
     compare = CompareItemButton()
     if compare.validate_on_submit():
-        items = Item.query.filter_by(Owner=None)
+        items = Item.query.all()
         return render_template('market-compare.html', items=items, item_id=compare.item_id.data, form=compare)
     
 @myapp.route('/comparing/<item1_id>&<item2_id>', methods = ['POST'])
@@ -232,5 +260,27 @@ def comparing(item1_id, item2_id):
     return render_template('comparing.html', price1=price1, item1=item1, item2=item2, purchase_form=purchase_form, add_to_cart = add_to_cart)
 
 
+@myapp.route('/seller_page)', methods = ['POST', 'GET'])
+def seller():
+    
+    seller_search = SellerPageForm()
+    checkout_form = purchaseItemForm()
+    if request.method == "POST":
+        checkout_item = request.form.get('checkout_item')
+        p_item_object = Item.query.filter_by(id=checkout_item).first()
+        if p_item_object:
+            if current_user.can_purchase(p_item_object):
+                p_item_object.buy(current_user)
+                flash(f"Congratulations! You purchased {p_item_object.name} for {p_item_object.price}$", category='success')
+            else:
+                flash(f"Unfortunately, you don't have enough money to purchase {p_item_object.name}!", category='danger')
 
+    if seller_search.validate_on_submit():
+
+        seller_user = User.query.filter_by(username = seller_search.name.data)
+        items = Item.query.filter_by(seller = seller_user.first().id)
+        
+        return render_template('seller.html', items = items, form = seller_search)
+    items = Item.query.filter_by(seller = current_user.id)
+    return render_template('seller.html', items = items, username = current_user, form = seller_search, checkout_form = checkout_form)
